server/requirements.txt
aiosmtplib==4.0.0 
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
DateTime==5.5
distro==1.9.0
dnspython==2.7.0
docx==0.2.4
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.6
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-genai==1.12.1
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
jiter==0.8.2
lxml==5.3.0
motor==3.6.1
openai==1.59.3
packaging==24.2
passlib==1.7.4
pillow==11.1.0

pluggy==1.5.0
proto-plus==1.25.0
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pyee==12.0.0
pymongo==4.9.2
pyparsing==3.2.0
PyPDF2==3.0.1
pytest==8.3.4
pytest-base-url==2.1.0
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-docx==1.1.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2024.2
requests==2.32.3
rsa==4.9
setuptools==75.6.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.41.3
text-unidecode==1.3
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
uvicorn==0.32.1
websockets==13.0.0
zipp==3.21.0
zope.interface==7.2
#redis
redis==6.2.0
hiredis==3.2.1
#ai-interviewer
groq==0.18.0
numpy==2.2.3
#speech api
# google-cloud-speech==2.30.0
google-cloud-texttospeech==2.24.0
google-cloud-storage
#rate-limiting
Deprecated==1.2.18
limits==4.4.1
slowapi==0.1.9
wrapt==1.17.2
#payments
dodopayments
#ocr
pytesseract==0.3.13
pymupdf==1.25.5
httpx==0.28.1
#logging
colorlog==6.9.0
#webhook signature verify Dodo
standardwebhooks==1.0.0
#ats nlp resume analysis
spacy==3.8.7
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
validators==0.35.0
#currency_details
countryinfo==0.1.2
currency-symbols==2.0.4